Revert "Deliver cookie also to NTP *server*, note about baking"
This reverts commit 40086ea5ab382bd8af6e376a15b39aedba7680e3.

We realise that we need not share cookies "live" between NTS-KE and NTP-server.
We need to share the parameters for encryption of the cookie, so that NTS-KE
can encrypt it before giving it to the client, and so NTP-server then can
decrypt it when upon receiving it from the client.